Move GetClusterZones() to e2enode
