+++ test/e2e/framework/skipper/BUILD
@@ -16,6 +16,7 @@ go_library(
         "//staging/src/k8s.io/client-go/dynamic:go_default_library",
         "//staging/src/k8s.io/client-go/kubernetes:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/framework/node:go_default_library",
         "//test/e2e/framework/ssh:go_default_library",
         "//vendor/github.com/onsi/ginkgo:go_default_library",
     ],
